Move timing functions into library
