+++ src/examples/cpc/cpc3-chi2NEQ.cpp
@@ -18,53 +18,6 @@ using namespace std;
 using namespace thermalfist;
 #endif
 
-// Time keeping
-// Windows
-#ifdef _WIN32
-#include <Windows.h>
-double get_wall_time() {
-  LARGE_INTEGER time, freq;
-  if (!QueryPerformanceFrequency(&freq)) {
-    //  Handle error
-    return 0;
-  }
-  if (!QueryPerformanceCounter(&time)) {
-    //  Handle error
-    return 0;
-  }
-  return (double)time.QuadPart / freq.QuadPart;
-}
-double get_cpu_time() {
-  FILETIME a, b, c, d;
-  if (GetProcessTimes(GetCurrentProcess(), &a, &b, &c, &d) != 0) {
-    //  Returns total user time.
-    //  Can be tweaked to include kernel times as well.
-    return
-      (double)(d.dwLowDateTime |
-      ((unsigned long long)d.dwHighDateTime << 32)) * 0.0000001;
-  }
-  else {
-    //  Handle error
-    return 0;
-  }
-}
-
-//  Posix/Linux
-#else
-#include <time.h>
-#include <sys/time.h>
-double get_wall_time() {
-  struct timeval time;
-  if (gettimeofday(&time, NULL)) {
-    //  Handle error
-    return 0;
-  }
-  return (double)time.tv_sec + (double)time.tv_usec * .000001;
-}
-double get_cpu_time() {
-  return (double)clock() / CLOCKS_PER_SEC;
-}
-#endif
 
 // Temperature dependence of the fit to ALICE 2.76 TeV data, 0-5% centrality, as in 1512.08046
 // Three variants of HRG model:
